feat(runtime): move app_related.go to internal/runtime (AS-154)
Rebased onto origin/main post-AS-150 merge (commit 1887860).

Migrates app_related.go (413 LOC) from internal/tui to the
platform-agnostic internal/runtime package per the Phase-05 boundary
contract (no bubbletea/lipgloss imports in runtime/).

- internal/runtime/related.go: HandleRelatedCheckStarted,
  BuildResourceCacheSnapshot, SnapshotCache, MissingFromCache,
  EnterChildForResource, BuildChildContextForResource,
  RelatedTitleSuffix, KindRelatedCheck, RelatedCheckStartedEvent.
  SnapshotCache uses maps.Copy (AS-209 squash; mapsloop lint).
- internal/tui/runtime_adapter_related.go: now hosts BOTH the
  AS-150 navigation adapter (handleRelatedNavigate, ChildView,
  relatedNavigateTasksToCmd) AND the AS-154 check adapter
  (handleRelatedCheckStarted, relatedCheckCmd). Navigation
  helpers updated to call runtime.RelatedTitleSuffix,
  runtime.EnterChildForResource, runtime.BuildChildContextForResource.
- internal/tui/related_helpers.go: BT-coupled helpers (newRelatedList,
  buildResourceCacheSnapshot wrapper, snapshotCache wrapper) that
  stay in the tui package.
- internal/tui/app_related.go: deleted (content migrated above).
- tests/unit/app_cancellation_test.go: static-scan path updated from
  app_related.go to runtime_adapter_related.go (goroutines now live
  in the merged adapter).

Rebase resolution preserves the net intent of both AS-150
(handler split into internal/runtime/handlers_related.go +
internal/tui/runtime_adapter_related.go) and AS-154 (related-check
handlers move to internal/runtime/related.go). The previously
separate adapter files collapse into a single
runtime_adapter_related.go because both adapters serve the same
architectural role (BT glue for runtime entry points).
